data/requirements.txt
streamlit==1.32.0
pandas==2.1.0
plotly==5.18.0
pillow==10.1.0
numpy==1.26.0